fixtures/ebfa/ebfa_fixtures_change_detect/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.30.0
google-cloud-pubsub
gcsfs==0.6.2
pandas==1.1.0
beautifulsoup4==4.9.3
soupsieve>=1.2.0
